Rename T3000 ts files
