--- src/lib/SDJS.js
@@ -1,9 +1,0 @@
-import Models from "./SDJS.Models";
-import Utils from "./SDJS.Utils";
-import App from "./SDJS.App";
-
-export default {
-  Models,
-  Utils,
-  App,
-};
