+++ src/lib/T3.ts
@@ -1,0 +1,10 @@
+// T3000
+import Models from "./T3.Models";
+import Utils from "./T3.Utils";
+import App from "./T3.App";
+
+export default {
+  Models,
+  Utils,
+  App,
+};
