--- src/lib/T3000.ts
@@ -1,10 +1,0 @@
-// T3000
-import Models from "./T3000.Models";
-import Utils from "./T3000.Utils";
-import App from "./T3000.App";
-
-export default {
-  Models,
-  Utils,
-  App,
-};
