+++ src/lib/T3JS.ts
@@ -1,0 +1,9 @@
+import Models from "./T3JS.Models";
+import Utils from "./T3JS.Utils";
+import App from "./T3JS.App";
+
+export default {
+  Models,
+  Utils,
+  App,
+};
